Emerging Infectious Network (EIN) Management of S. aureus Bacteremia survey data

Responses to survey questions about infectious disease consultation for staphylococcus aureus bacteremia (SAB) presented to medical personnel who are members of EIN.

Usage

data(bacteremia)

Format

bacteremia

A list with 15 objects that each have a length of 669.

q1

A character vector that lists the responses to question 1 case study.

q2

A character vector that lists the responses to question 2 case study.

q3

A character vector that lists the responses to preferred choice of valve imaging in a patient with MSSA bacteremia.

q4

A character vector that lists the responses on how to treat a clinically stable patient that has MRSA bacteremia.

q5

A character vector that lists the responses on antibiotic treatment strategy for a patient with MRSA endocarditis on day 6 of vancomycin with persistent positive blood cultures.

q6

A character vector that lists the responses to the treatment course for a 35 y/o previously healthy man with left-sided MSSA endocarditis but no meningitis or CNS disease.

q7_bc

A character vector that lists the responses to how often repeat blood cultures from 24 hours to 48 hours until negative is used when evaluating a patient with SAB.

q7_echo

A character vector that lists the responses to how often a TTE and/or TEE is used when evaluating a patient with SAB.

q7_spine

A character vector that lists the responses to how often a MRI of the spine is used when evaluating a patient with SAB.

q7_abd

A character vector that lists the responses to how often a CT of the abdominal/pelvic area is used when evaluating a patient with SAB.

q7_eye

A character vector that lists the responses to how often a dilated eye exam is used when evaluating a patient with SAB.

q9

A character vector that lists the responses to dosage of daptomycin for the treatment of MRSA bacteremia.

q10

A character vector that lists the responses to management strategy and decision regarding anticoagulation on patients with MSSA bacteremia and a deep vein thrombus at the site of a recently removed PICC line.

q8

A data.frame of 7 character variables that indicate whether response resulted in the selection of the select all apply response item. The responses are focused on what conditions would increase the treatment duration from 2 weeks to 4-6 weeks assuming that a TTE and/or TEE are negative in a patient with MRSA bacteremia. If the response item was selected a value of Yes is reported while No indicates the condition was not selected.

wave

A numeric vector that lists of the data collect period when the responses were collected during.

Source

S. aureus Bacteremia

Details

Questions 1, 2, 3, 5, and 6 are clinical case study questions, where the remaining questions ask about different treatment methodologies. Questions 1 and 2 provide a medical history in the case study, which is excluded in the descriptions of the responses for q1 and q2a. These questions are asking what would be the treatment strategy of an infectious disease physician given the full medical history of the patients.

Examples

data(bacteremia)
str(bacteremia)
#> List of 15
#>  $ q1      : chr [1:669] "IV vancomycin for 14 days" "Oral antibiotics for 14 days" "IV vancomycin for 14 days" "Oral antibiotics for 14 days" ...
#>  $ q2      : chr [1:669] "Transition to oral antibiotics" "Initial cultures likely contaminant" "Continue IV antibiotics" "Transition to oral antibiotics" ...
#>  $ q3      : chr [1:669] "TTE for all, but if negative TEE only if ..." "TTE for all, but if negative TEE only if ..." "TTE for all, and if negative TEE for all" "TTE for all, but if negative TEE only if ..." ...
#>  $ q4      : chr [1:669] "Treat with other antibiotics" "Treat with other antibiotics" "Treat with other antibiotics" "Treat with other antibiotics" ...
#>  $ q5      : chr [1:669] "Change to other antibiotic" "Other treatment" "Change to daptomycin and add MRSA" "Change to other antibiotic" ...
#>  $ q5      : chr [1:669] "Cefazolin or Nafcillin" "Cefazolin" "Cefazolin or Nafcillin" "Other" ...
#>  $ q7_bc   : chr [1:669] "Almost always" "Almost always" "Almost always" "Sometimes" ...
#>  $ q7_echo : chr [1:669] "Almost always" "Sometimes" "Almost always" "Sometimes" ...
#>  $ q7_spine: chr [1:669] "Rarely" "Sometimes" "Sometimes" "Rarely" ...
#>  $ q7_abd  : chr [1:669] "Rarely" "Rarely" "Sometimes" "Rarely" ...
#>  $ q7_eye  : chr [1:669] "Rarely" "Sometimes" "Rarely" "Rarely" ...
#>  $ q9      : chr [1:669] "8 mg/kg IV Q24hrs" "8 mg/kg IV Q24hrs" "8 mg/kg IV Q24hrs" "8 mg/kg IV Q24hrs" ...
#>  $ q10     : chr [1:669] "Six weeks with anticoagulation" "Two weeks with no anticoagulation" "Two weeks with anticoagulation" "Four weeks with no anticoagulation" ...
#>  $ q8      :'data.frame':	669 obs. of  7 variables:
#>   ..$ repeat_bc : chr [1:669] "Yes" "Yes" "Yes" "Yes" ...
#>   ..$ fevers    : chr [1:669] "Yes" "No" "No" "No" ...
#>   ..$ diabetes  : chr [1:669] "No" "No" "No" "No" ...
#>   ..$ immuno    : chr [1:669] "No" "No" "No" "No" ...
#>   ..$ prosthetic: chr [1:669] "Yes" "No" "Yes" "Yes" ...
#>   ..$ community : chr [1:669] "No" "No" "No" "No" ...
#>   ..$ other     : chr [1:669] "No" "No" "No" "No" ...
#>  $ wave    : num [1:669] 2 1 1 1 1 1 2 2 1 3 ...