Emerging Infectious Network (EIN) survey data

Responses to survey questions about advanced molecular diagnostic testing for infectious diseases presented to medical personnel who are members of EIN.

Usage

data(ein)

Format

ein

A data.frame with 643 observations and 3 variables.

wave

The data collect period the responses were collected during.

q1

The responses to the types of advanced molecular diagnostic tests that have been ordered or used. This is an example of responses based on select that all apply data collection process. Possible values are Broad range 16S rRNA gene sequencing (Broadrange), Whole genome sequencing (Wholegenome), Metagenomic next-generation sequencing (MNGS), None of the above (None), and Other (Other), where the data values are listed within the parentheses. When multiple responses are present the data values are separated by "|".

q2

The responses to how often within the past two years advanced molecular diagnostic testing was performed. Possible values are Not at all, Once, Rarely (quarterly or less often), Sometimes (monthly), and Often (weekly). This is an example of where the responses collected come from the multinomial distribution.

Source

Difficult Diagnoses

Examples

data(ein)
str(ein)
#> 'data.frame':	643 obs. of  3 variables:
#>  $ wave: num  2 3 3 2 2 3 3 3 2 3 ...
#>  $ q1  : chr  "Broadrange|Wholegenome|MNGS" "Broadrange|Wholegenome|MNGS" "Broadrange|MNGS" "Broadrange" ...
#>  $ q2  : Factor w/ 5 levels "Not at all","Once",..: 4 5 4 2 3 3 3 3 4 4 ...