satpt: Saturation Point Analysis for Sample Proportions

Identify the saturation point of multinomial responses from a survey using standard errors of sample proportions.

References

Boonstra DE, Cavanaugh JE, Miller AR, and Polgreen PM. (2025) Saturation point analysis for surveys using standard errors of sample proportions. BMC Medical Informatics and Decision Making, Vol.: #-#.

See also

Author

Maintainer: D. Erik Boonstra daniel-boonstra@uiowa.edu (ORCID)

Authors:

Joseph E. Cavanaugh (ORCID)