Simulation of multinomial responses

Simulating the data of multinomial responses collected during a survey.

Usage

simulate(n, size, prob, categories = NULL)

Arguments

n

Integer specifying the number of random samples to draw.

size

Vector of integer values, say \(N_i\), specifying the sample sizes for \(i = 1, \ldots , I\) data collection periods.

prob

Numeric non-negative vector or matrix specifying the probabilities for the K response categories in the \(I\) data collection periods.

categories

Vector of values specifying the K category labels to be represented. Default is NULL.

Value

A data.frame with \(N = \sum_{i}N_{i}\) rows and n + 1 variables, which are described below.

period

Data collection period defined by the length of size

responses1 ... n

Simulated multinomial responses with either the values defined by categories or Category k, for \(k = 1, \ldots , K\)

Details

Running in the backend is stats::rmultinom(). So, see the documentation for stats::rmultinom()for more information on how the multinomial responses are generated.

The arguments size and prob are directly connect when it comes to generating multinomial responses for multiple data collection periods. prob may either be a vector or matrix specifying probabilities for the K response categories. When size is specified to generate data for multiple data collection periods, then prob must be specified as a numeric vector of length length(size) times K or numeric matrix with length(size) rows and K columns, where each row of the matrix specify the K probabilities of each data collection period.

When categories is not defined (categories = NULL) simulate will create default lables of Category k, for \(k = 1, \ldots , K\).

See also

Examples

# Creating 10 simulated data sets of with a sample size of 100, where there
# are two possible response categories with a 50% chance of being selected.
simulate(n = 10, size = 100, prob = c(0.5, 0.5))
#>     period responses1 responses2 responses3 responses4 responses5 responses6
#> 1        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 2        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 3        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 4        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 5        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 6        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 7        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 8        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 9        1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 10       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 11       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 12       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 13       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 14       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 15       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 16       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 17       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 18       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 19       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 20       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 21       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 22       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 23       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 24       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 25       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 26       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 27       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 28       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 29       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 30       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 31       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 32       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 33       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 34       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 35       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 36       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 37       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 38       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 39       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 40       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 41       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 42       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 43       1 Category 1 Category 1 Category 1 Category 1 Category 1 Category 1
#> 44       1 Category 1 Category 1 Category 1 Category 2 Category 1 Category 1
#> 45       1 Category 2 Category 1 Category 1 Category 2 Category 1 Category 1
#> 46       1 Category 2 Category 1 Category 1 Category 2 Category 1 Category 1
#> 47       1 Category 2 Category 1 Category 2 Category 2 Category 1 Category 1
#> 48       1 Category 2 Category 1 Category 2 Category 2 Category 1 Category 1
#> 49       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 50       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 51       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 52       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 53       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 54       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 55       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 56       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 1
#> 57       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 58       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 59       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 60       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 61       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 62       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 63       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 64       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 65       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 66       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 67       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 68       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 69       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 70       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 71       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 72       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 73       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 74       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 75       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 76       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 77       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 78       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 79       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 80       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 81       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 82       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 83       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 84       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 85       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 86       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 87       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 88       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 89       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 90       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 91       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 92       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 93       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 94       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 95       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 96       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 97       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 98       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 99       1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#> 100      1 Category 2 Category 2 Category 2 Category 2 Category 2 Category 2
#>     responses7 responses8 responses9 responses10
#> 1   Category 1 Category 1 Category 1  Category 1
#> 2   Category 1 Category 1 Category 1  Category 1
#> 3   Category 1 Category 1 Category 1  Category 1
#> 4   Category 1 Category 1 Category 1  Category 1
#> 5   Category 1 Category 1 Category 1  Category 1
#> 6   Category 1 Category 1 Category 1  Category 1
#> 7   Category 1 Category 1 Category 1  Category 1
#> 8   Category 1 Category 1 Category 1  Category 1
#> 9   Category 1 Category 1 Category 1  Category 1
#> 10  Category 1 Category 1 Category 1  Category 1
#> 11  Category 1 Category 1 Category 1  Category 1
#> 12  Category 1 Category 1 Category 1  Category 1
#> 13  Category 1 Category 1 Category 1  Category 1
#> 14  Category 1 Category 1 Category 1  Category 1
#> 15  Category 1 Category 1 Category 1  Category 1
#> 16  Category 1 Category 1 Category 1  Category 1
#> 17  Category 1 Category 1 Category 1  Category 1
#> 18  Category 1 Category 1 Category 1  Category 1
#> 19  Category 1 Category 1 Category 1  Category 1
#> 20  Category 1 Category 1 Category 1  Category 1
#> 21  Category 1 Category 1 Category 1  Category 1
#> 22  Category 1 Category 1 Category 1  Category 1
#> 23  Category 1 Category 1 Category 1  Category 1
#> 24  Category 1 Category 1 Category 1  Category 1
#> 25  Category 1 Category 1 Category 1  Category 1
#> 26  Category 1 Category 1 Category 1  Category 1
#> 27  Category 1 Category 1 Category 1  Category 1
#> 28  Category 1 Category 1 Category 1  Category 1
#> 29  Category 1 Category 1 Category 1  Category 1
#> 30  Category 1 Category 1 Category 1  Category 1
#> 31  Category 1 Category 1 Category 1  Category 1
#> 32  Category 1 Category 1 Category 1  Category 1
#> 33  Category 1 Category 1 Category 1  Category 1
#> 34  Category 1 Category 1 Category 1  Category 1
#> 35  Category 1 Category 1 Category 1  Category 1
#> 36  Category 1 Category 1 Category 1  Category 1
#> 37  Category 1 Category 1 Category 1  Category 1
#> 38  Category 1 Category 1 Category 1  Category 1
#> 39  Category 1 Category 1 Category 1  Category 1
#> 40  Category 1 Category 1 Category 1  Category 1
#> 41  Category 1 Category 1 Category 1  Category 1
#> 42  Category 1 Category 1 Category 1  Category 1
#> 43  Category 1 Category 1 Category 1  Category 1
#> 44  Category 1 Category 1 Category 1  Category 1
#> 45  Category 1 Category 1 Category 1  Category 1
#> 46  Category 1 Category 1 Category 1  Category 1
#> 47  Category 1 Category 1 Category 1  Category 1
#> 48  Category 1 Category 1 Category 1  Category 1
#> 49  Category 1 Category 1 Category 1  Category 1
#> 50  Category 1 Category 2 Category 1  Category 1
#> 51  Category 1 Category 2 Category 2  Category 1
#> 52  Category 1 Category 2 Category 2  Category 1
#> 53  Category 1 Category 2 Category 2  Category 1
#> 54  Category 1 Category 2 Category 2  Category 2
#> 55  Category 1 Category 2 Category 2  Category 2
#> 56  Category 1 Category 2 Category 2  Category 2
#> 57  Category 1 Category 2 Category 2  Category 2
#> 58  Category 2 Category 2 Category 2  Category 2
#> 59  Category 2 Category 2 Category 2  Category 2
#> 60  Category 2 Category 2 Category 2  Category 2
#> 61  Category 2 Category 2 Category 2  Category 2
#> 62  Category 2 Category 2 Category 2  Category 2
#> 63  Category 2 Category 2 Category 2  Category 2
#> 64  Category 2 Category 2 Category 2  Category 2
#> 65  Category 2 Category 2 Category 2  Category 2
#> 66  Category 2 Category 2 Category 2  Category 2
#> 67  Category 2 Category 2 Category 2  Category 2
#> 68  Category 2 Category 2 Category 2  Category 2
#> 69  Category 2 Category 2 Category 2  Category 2
#> 70  Category 2 Category 2 Category 2  Category 2
#> 71  Category 2 Category 2 Category 2  Category 2
#> 72  Category 2 Category 2 Category 2  Category 2
#> 73  Category 2 Category 2 Category 2  Category 2
#> 74  Category 2 Category 2 Category 2  Category 2
#> 75  Category 2 Category 2 Category 2  Category 2
#> 76  Category 2 Category 2 Category 2  Category 2
#> 77  Category 2 Category 2 Category 2  Category 2
#> 78  Category 2 Category 2 Category 2  Category 2
#> 79  Category 2 Category 2 Category 2  Category 2
#> 80  Category 2 Category 2 Category 2  Category 2
#> 81  Category 2 Category 2 Category 2  Category 2
#> 82  Category 2 Category 2 Category 2  Category 2
#> 83  Category 2 Category 2 Category 2  Category 2
#> 84  Category 2 Category 2 Category 2  Category 2
#> 85  Category 2 Category 2 Category 2  Category 2
#> 86  Category 2 Category 2 Category 2  Category 2
#> 87  Category 2 Category 2 Category 2  Category 2
#> 88  Category 2 Category 2 Category 2  Category 2
#> 89  Category 2 Category 2 Category 2  Category 2
#> 90  Category 2 Category 2 Category 2  Category 2
#> 91  Category 2 Category 2 Category 2  Category 2
#> 92  Category 2 Category 2 Category 2  Category 2
#> 93  Category 2 Category 2 Category 2  Category 2
#> 94  Category 2 Category 2 Category 2  Category 2
#> 95  Category 2 Category 2 Category 2  Category 2
#> 96  Category 2 Category 2 Category 2  Category 2
#> 97  Category 2 Category 2 Category 2  Category 2
#> 98  Category 2 Category 2 Category 2  Category 2
#> 99  Category 2 Category 2 Category 2  Category 2
#> 100 Category 2 Category 2 Category 2  Category 2

# Creating 1 simulated data set for two data collection periods, where there
# are three possible response categories that are labeled.
prob <- matrix(
 data = c(0.4, 0.4, 0.2, 0.1, 0.1, 0.8),
 nrow = 2, ncol = 3, byrow = TRUE
)
categories <- LETTERS[1:3]
simulate(n = 1, size = c(200, 100), prob = prob, categories = categories)
#>     period responses1
#> 1        1          A
#> 2        1          A
#> 3        1          A
#> 4        1          A
#> 5        1          A
#> 6        1          A
#> 7        1          A
#> 8        1          A
#> 9        1          A
#> 10       1          A
#> 11       1          A
#> 12       1          A
#> 13       1          A
#> 14       1          A
#> 15       1          A
#> 16       1          A
#> 17       1          A
#> 18       1          A
#> 19       1          A
#> 20       1          A
#> 21       1          A
#> 22       1          A
#> 23       1          A
#> 24       1          A
#> 25       1          A
#> 26       1          A
#> 27       1          A
#> 28       1          A
#> 29       1          A
#> 30       1          A
#> 31       1          A
#> 32       1          A
#> 33       1          A
#> 34       1          A
#> 35       1          A
#> 36       1          A
#> 37       1          A
#> 38       1          A
#> 39       1          A
#> 40       1          A
#> 41       1          A
#> 42       1          A
#> 43       1          A
#> 44       1          A
#> 45       1          A
#> 46       1          A
#> 47       1          A
#> 48       1          A
#> 49       1          A
#> 50       1          A
#> 51       1          A
#> 52       1          A
#> 53       1          A
#> 54       1          A
#> 55       1          A
#> 56       1          A
#> 57       1          A
#> 58       1          A
#> 59       1          A
#> 60       1          A
#> 61       1          A
#> 62       1          A
#> 63       1          A
#> 64       1          A
#> 65       1          A
#> 66       1          A
#> 67       1          A
#> 68       1          A
#> 69       1          A
#> 70       1          A
#> 71       1          A
#> 72       1          A
#> 73       1          A
#> 74       1          A
#> 75       1          A
#> 76       1          B
#> 77       1          B
#> 78       1          B
#> 79       1          B
#> 80       1          B
#> 81       1          B
#> 82       1          B
#> 83       1          B
#> 84       1          B
#> 85       1          B
#> 86       1          B
#> 87       1          B
#> 88       1          B
#> 89       1          B
#> 90       1          B
#> 91       1          B
#> 92       1          B
#> 93       1          B
#> 94       1          B
#> 95       1          B
#> 96       1          B
#> 97       1          B
#> 98       1          B
#> 99       1          B
#> 100      1          B
#> 101      1          B
#> 102      1          B
#> 103      1          B
#> 104      1          B
#> 105      1          B
#> 106      1          B
#> 107      1          B
#> 108      1          B
#> 109      1          B
#> 110      1          B
#> 111      1          B
#> 112      1          B
#> 113      1          B
#> 114      1          B
#> 115      1          B
#> 116      1          B
#> 117      1          B
#> 118      1          B
#> 119      1          B
#> 120      1          B
#> 121      1          B
#> 122      1          B
#> 123      1          B
#> 124      1          B
#> 125      1          B
#> 126      1          B
#> 127      1          B
#> 128      1          B
#> 129      1          B
#> 130      1          B
#> 131      1          B
#> 132      1          B
#> 133      1          B
#> 134      1          B
#> 135      1          B
#> 136      1          B
#> 137      1          B
#> 138      1          B
#> 139      1          B
#> 140      1          B
#> 141      1          B
#> 142      1          B
#> 143      1          B
#> 144      1          B
#> 145      1          B
#> 146      1          B
#> 147      1          B
#> 148      1          B
#> 149      1          B
#> 150      1          B
#> 151      1          B
#> 152      1          B
#> 153      1          B
#> 154      1          B
#> 155      1          B
#> 156      1          B
#> 157      1          B
#> 158      1          B
#> 159      1          B
#> 160      1          C
#> 161      1          C
#> 162      1          C
#> 163      1          C
#> 164      1          C
#> 165      1          C
#> 166      1          C
#> 167      1          C
#> 168      1          C
#> 169      1          C
#> 170      1          C
#> 171      1          C
#> 172      1          C
#> 173      1          C
#> 174      1          C
#> 175      1          C
#> 176      1          C
#> 177      1          C
#> 178      1          C
#> 179      1          C
#> 180      1          C
#> 181      1          C
#> 182      1          C
#> 183      1          C
#> 184      1          C
#> 185      1          C
#> 186      1          C
#> 187      1          C
#> 188      1          C
#> 189      1          C
#> 190      1          C
#> 191      1          C
#> 192      1          C
#> 193      1          C
#> 194      1          C
#> 195      1          C
#> 196      1          C
#> 197      1          C
#> 198      1          C
#> 199      1          C
#> 200      1          C
#> 201      2          A
#> 202      2          A
#> 203      2          A
#> 204      2          A
#> 205      2          A
#> 206      2          A
#> 207      2          A
#> 208      2          A
#> 209      2          B
#> 210      2          B
#> 211      2          B
#> 212      2          B
#> 213      2          B
#> 214      2          B
#> 215      2          B
#> 216      2          B
#> 217      2          B
#> 218      2          B
#> 219      2          B
#> 220      2          C
#> 221      2          C
#> 222      2          C
#> 223      2          C
#> 224      2          C
#> 225      2          C
#> 226      2          C
#> 227      2          C
#> 228      2          C
#> 229      2          C
#> 230      2          C
#> 231      2          C
#> 232      2          C
#> 233      2          C
#> 234      2          C
#> 235      2          C
#> 236      2          C
#> 237      2          C
#> 238      2          C
#> 239      2          C
#> 240      2          C
#> 241      2          C
#> 242      2          C
#> 243      2          C
#> 244      2          C
#> 245      2          C
#> 246      2          C
#> 247      2          C
#> 248      2          C
#> 249      2          C
#> 250      2          C
#> 251      2          C
#> 252      2          C
#> 253      2          C
#> 254      2          C
#> 255      2          C
#> 256      2          C
#> 257      2          C
#> 258      2          C
#> 259      2          C
#> 260      2          C
#> 261      2          C
#> 262      2          C
#> 263      2          C
#> 264      2          C
#> 265      2          C
#> 266      2          C
#> 267      2          C
#> 268      2          C
#> 269      2          C
#> 270      2          C
#> 271      2          C
#> 272      2          C
#> 273      2          C
#> 274      2          C
#> 275      2          C
#> 276      2          C
#> 277      2          C
#> 278      2          C
#> 279      2          C
#> 280      2          C
#> 281      2          C
#> 282      2          C
#> 283      2          C
#> 284      2          C
#> 285      2          C
#> 286      2          C
#> 287      2          C
#> 288      2          C
#> 289      2          C
#> 290      2          C
#> 291      2          C
#> 292      2          C
#> 293      2          C
#> 294      2          C
#> 295      2          C
#> 296      2          C
#> 297      2          C
#> 298      2          C
#> 299      2          C
#> 300      2          C