Saturation point analysis

Saturation point analysis of multinomial responses from a survey using standard errors of the sample proportions for the responses.

Usage

satpt(
  y,
  by = NULL,
  exclude = c(NA, NaN),
  alpha = 0.05,
  threshold = 0.05,
  dimnames = NULL
)

Arguments

y

Vector of values to which determine if saturation has been achieved.

by

Vector of values which indicate when y was collected. Default is NULL.

exclude

Vector of values that should be excluded in y and by. Generally, this should be used to denote missing values. Default is NA and NaN.

alpha

Significance level for Pearson's \(\chi^2\) test for independence. Default is 0.05.

threshold

Saturation threshold applied to the standard errors of the sample proportions. Default is 0.05.

dimnames

Character vector of names for y and by when displaying the contingency table, sample proportions, and standard error matrices. When dimnames is an unnamed vector the first entry should be name of y variable and the second entry should be name of by variable. If dimnames is a named vector, then order of the values do NOT matter, as long the elements are named with "y" and "by". Default is NULL.

Value

An object with S3 class "satpt" containing 11 elements.

threshold

Saturation threshold applied to the standard errors of thesample proportions.

saturation

A logical value indicating whether all response categories have achieved saturation given the defined threshold. The value of TRUE indicates that saturation has been achieved while a value of FALSE indicates that saturation was not achieved and more data is needed to achieve saturation.

counts

A matrix containing the observed cell counts of the contigency table created by y and by if provided.

phat

A matrix containing the row-wise sample proportions for the observed contigency table (counts).

se

A matrix containing the standard errors for the calculated sample proportions (phat).

pooled_se

A logical value indicating whether pooled standard errors were calculated due to the presence of response bias.

alpha

Significance level for Pearson's \(\chi^2\) test for independence.

test

A htest object produced by stats::chisq.test() containing the results from the \(\chi^2\) test for independence.

N

Total number of observations with a response provided.

total

A data.frame with 5 variables describing the overall collected sample. The categories variable provides the unique categories listed in y. While counts, phat, and se provide the overall cell counts, sample proportions, and standard errors for the categories, respectively. The standard errors reported for the overall sample proportions are calculated based on the presence of response bias, which is detailed above. Finally, the saturation variable is a logical variable denoting whether saturation was achieved for each individual category given the saturation threshold defined.

hindex

Heterogeneity index for the sample proportions calculated by mean absolute deviation.

Details

The by argument should be specified when the responses collected in y are a by-product of the data collection mechanism defined by the by variable. When there is no a priori data collection mechanism defined, the by argument should not be defined. Generally, when the data is collected randomly or collected during the first data collection period, there is "no a priori data collection mechanism."

Functionality of satpt() depends on the limited use of stats::ftable(). More specifically, y, by, and exclude are directly used with stats::ftable() to create the contingency table of the collected data. The contingency table created by stats::ftable() is converted to a matrix for easier use with other functions. When the dimnames argument is used, it manipulates the attributes of the created ftable, which impacts the dimension names of the resulting matrix.

Specification of the by argument automatically calls for a Pearson's \(\chi^2\) test for independence and the heterogeneity index of the sample proportions to be calculated. The \(\chi^2\) test for independence is conducted to determine if response bias is present within the responses in y are due to the data collection periods (by). When response bias is present the pooled standard errors of the overall sample proportions for each response category is reported, the pooled standard errors account for the response bias. The heterogeneity index is defined as the mean absoluted difference of the sample proportions for each response category within each data collection period (by) relative to the overall sample proportions for each response category. This index reflects the average deviation of the data collection period proportions from the overall sample proportions. Smaller values indicate the sample proportions of the data collection periods are less dissimilar. This measure is of importance when response bias is present. When by is not specified the test for independence will not be conducted and the heterogeneity index will not be calculated. satpt assumes response bias is only possible when by is specified. Thus, there is no need to check for response bias when by is not specified.

Note

The returned satpt object will contain NULL values for test and hindex when by is not specified. This is done because satpt assumes by is only specified when the data is collected in intervals.

See also

Examples

data(ein)

# Assuming response bias is not a possiblity
satpt::satpt(y = ein$q2)
#> Saturation achieved?  Yes 
#> Overall Sample Proportions and Standard Errors
#> ==============================================
#>             y
#> Statistics   Not at all   Once Rarely Sometimes  Often
#>   Proportion     0.2531 0.0375 0.3688    0.2656 0.0750
#>   SE             0.0172 0.0075 0.0191    0.0175 0.0104

# Examining saturation given data collected at different times and
# response bias is possible. For this example, response bias is not present,
# so the standard errors will be the same.
satpt::satpt(y = ein$q2, by = ein$wave)
#> Saturation achieved?  Yes 
#> Overall Sample Proportions and Standard Errors
#> ==============================================
#>             y
#> Statistics   Not at all   Once Rarely Sometimes  Often
#>   Proportion     0.2531 0.0375 0.3688    0.2656 0.0750
#>   SE             0.0172 0.0075 0.0191    0.0175 0.0104

# Creating an example, where response bias is present.

## Simulating data
prob <- matrix(
  data = c(0.4, 0.4, 0.2, 0.1, 0.1, 0.8),
  nrow = 2, ncol = 3, byrow = TRUE
)
catg <- LETTERS[1:3]
set.seed(123)
dat <- satpt::simulate(
  n = 1, size = c(200, 100), prob = prob, categories = catg
)

## Determining saturation with response bias
res <- satpt::satpt(y = dat$responses1, by = dat$period)
summary(res)
#> 
#> Saturation point analysis of sample proportions
#> ===============================================
#> 
#> Saturation achieved? Yes
#> Saturation threshold of 0.05
#> Responses collected from a sample size of 300
#> 
#> Data interval and overall sample proportions
#> ============================================
#>          y
#> by             A      B      C
#>   1       0.3850 0.4050 0.2100
#>   2       0.1500 0.0500 0.8000
#>   Overall 0.3067 0.2867 0.4067
#> 
#> 
#> Data interval and overall standard errors
#> =========================================
#>          y
#> by             A      B      C
#>   1       0.0344 0.0347 0.0288
#>   2       0.0357 0.0218 0.0400
#>   Overall 0.0258 0.0243 0.0234
#> 
#> Pooled standard errors?  Yes 
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  y given by
#> X-squared = 98.379, df = 2, p-value < 2.2e-16
#> 
#> Response bias present? Yes
#> Significance level: 0.05
#> 
#> 
#> Heterogeneity index
#> ====================
#>  Categories  Index
#>           A 0.1175
#>           B 0.1775
#>           C 0.2950